modify merge conflict repo
* Beacon delete
* merge conflict modify
