+++ app/src/main/AndroidManifest.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
     package="io.kong.incheon.nfc_check">
 
     <application
@@ -9,8 +10,6 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-<<<<<<< HEAD
-
         <activity android:name=".activity.MainActivity"/>
         <activity android:name=".activity.TimeTableActivity">
 
@@ -21,11 +20,10 @@
             </intent-filter>
         </activity>
         <activity android:name=".activity.SubjectActivity"/>
-=======
 
-        <activity android:name=".SignupActivity" />
-        <activity android:name=".MainActivity"/>
->>>>>>> 180802_NfcTagUI
+        <activity android:name=".activity.SignupActivity" />
+        <activity android:name=".activity.MainActivity"
+            tools:ignore="DuplicateActivity" />
 
     </application>
