--- app/src/main/java/io/kong/incheon/nfc_check/Beacon.java
@@ -1,68 +1,0 @@
-package io.kong.incheon.nfc_check;
-
-import android.app.Application;
-import android.app.Notification;
-import android.app.NotificationManager;
-import android.app.PendingIntent;
-import android.content.Context;
-import android.content.Intent;
-
-import com.estimote.sdk.BeaconManager;
-import com.estimote.sdk.Region;
-
-import java.util.List;
-import java.util.UUID;
-
-public class Beacon extends Application{
-
-    private BeaconManager beaconManager;
-
-    @Override
-    public void onCreate() {
-        super.onCreate();
-        beaconManager = new BeaconManager(getApplicationContext());
-
-        //애플리케이션 설치가 끝나면 비콘 모니터링 서비스가 시작된다.
-        //어플리케이션을 종료하더라도 서비스가 계속 실행된다.
-        beaconManager.connect(new BeaconManager.ServiceReadyCallback() {
-            @Override
-            public void onServiceReady() {
-                beaconManager.startMonitoring(new Region(
-                        "monitored region", UUID.fromString(""), 0, 0));
-                //본인이 연결할 비콘 아이디와 코드를 써야함
-                //비콘의 아이디 값을 넣는 것이다.
-            }
-        });
-
-        beaconManager.setMonitoringListener(new BeaconManager.MonitoringListener() {
-            @Override
-            public void onEnteredRegion(Region region, List<com.estimote.sdk.Beacon> list) {
-                showNotification("들어옴","비콘 연결됨"+list.get(0).getRssi());
-
-            }
-
-            @Override
-            public void onExitedRegion(Region region) {
-                showNotification("나감","비콘 연결 끊김");
-            }
-        });
-    }
-
-    /* notification으로 비콘이 연결 되었거나 끊김을 알린다. */
-    public void showNotification(String title, String message){
-        Intent notifyIntent = new Intent(this, BeaconActivity.class);
-        notifyIntent.setFlags(Intent.FLAG_ACTIVITY_SINGLE_TOP);
-        PendingIntent pendingIntent = PendingIntent.getActivities(this,0,
-                new Intent[]{notifyIntent}, PendingIntent.FLAG_UPDATE_CURRENT);
-        Notification notification = new Notification.Builder(this)
-               .setContentTitle(title)
-                .setContentText(message)
-                .setAutoCancel(true)
-                .setContentIntent(pendingIntent)
-                .build();
-        notification.defaults |= Notification.DEFAULT_SOUND;
-        NotificationManager notificationManager =
-                (NotificationManager) getSystemService(Context.NOTIFICATION_SERVICE);
-        notificationManager.notify(1,notification);
-    }
-}
